добавлено: очистка группы, редактирование группы исправлено: загрузка списка групп, удаление группы, выбор группы

+++ ajax.php
@@ -7,20 +7,57 @@ if(isset($_POST['dell_group'])){
 		$id = htmlspecialchars($_POST['id']);
 		$pin = htmlspecialchars($_POST['pin']);
 		if($mysqli->query("DELETE FROM `grups` WHERE `id_grup`='$id' AND `pin`='$pin'")){
-			if($mysqli->query("DELETE FROM `raspis` WHERE `id_grup`='$id'")){
-			echo "Ok";
+			if($mysqli->affected_rows == '1'){
+				if($mysqli->query("DELETE FROM `raspis` WHERE `id_grup`='$id'"))
+					echo "Ok";
+				else
+					echo "Ошибка 3";
+			}else
+				echo "Ошибка 2";
 		}else
-			echo "Ошибка sql_2";
-		}
-		else
-			echo "Ошибка sql_1";
+			echo "Ошибка 1";
 	}
 }
-if(isset($_POST['bod_list_droup'])){
+if(isset($_POST['body_list_group'])){
 	if(isset($_SESSION['mast']) && $_SESSION['mast'] == $maspar){
 		echo "Ok".load_table_group($mysqli);
 	}else{
-		echo "error";
+		echo "Ошибка -1";
+	}
+}
+if(isset($_POST['clear_group'])){
+	if(isset($_POST['id']) && isset($_POST['pin'])){
+		$id = htmlspecialchars($_POST['id']);
+		$pin = htmlspecialchars($_POST['pin']);
+		if($result_c = $mysqli->query("SELECT * FROM `grups` WHERE `id_grup`='$id' AND `pin`='$pin'")){
+			if($result_c->num_rows == '1'){
+				if($mysqli->query("DELETE FROM `raspis` WHERE `id_grup`='$id'"))
+					echo "Ok";
+				else
+					echo "Ошибка 3";
+			}else
+				echo "Ошибка 2";
+		}else
+			echo "Ошибка 1";
+	}
+}
+if(isset($_POST['edit_group'])){
+	if(isset($_POST['id']) && isset($_POST['pin']) && isset($_POST['pin_old']) && isset($_POST['name']) && isset($_POST['date'])){
+		$id = htmlspecialchars($_POST['id']);
+		$pin = htmlspecialchars($_POST['pin']);
+		$pin_old = htmlspecialchars($_POST['pin_old']);
+		$name = htmlspecialchars($_POST['name']);
+		$date = htmlspecialchars($_POST['date']);
+		if($result_c = $mysqli->query("SELECT * FROM `grups` WHERE `id_grup`='$id' AND `pin`='$pin_old'")){
+			if($result_c->num_rows == '1'){
+				if($mysqli->query("UPDATE `grups` SET `name`='$name',`start`='$date',`pin`='$pin' WHERE `id_grup`='$id'"))
+					echo "Ok";
+				else
+					echo "Ошибка 3";
+			}else
+				echo "Ошибка 2";
+		}else
+			echo "Ошибка 1";
 	}
 }
 ?>
